删除容器前检查容器是否存在

+++ src/main/java/io/metersphere/service/JmeterOperateService.java
@@ -77,7 +77,9 @@ public class JmeterOperateService {
                             try {
                                 FileUtils.forceDelete(new File(filePath));
                                 LogUtil.info("Remove dir completed.");
-                                DockerClientService.removeContainer(dockerClient, containerId);
+                                if (DockerClientService.existContainer(dockerClient, containerId) > 0) {
+                                    DockerClientService.removeContainer(dockerClient, containerId);
+                                }
                                 LogUtil.info("Remove container completed: " + containerId);
                             } catch (IOException e) {
                                 LogUtil.error("Remove dir error: ", e);
